bot/requirements.txt
python-telegram-bot==21.10
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
